trying to implement acl

+++ homework_2/src/main/java/com/study/spring/library/service/AuthorServiceImpl.java
@@ -1,10 +1,13 @@
 package com.study.spring.library.service;
 
 import com.study.spring.library.dao.AuthorDao;
+import com.study.spring.library.dao.BookDao;
 import com.study.spring.library.domain.Author;
+import com.study.spring.library.exceptions.ConsistencyException;
 import com.study.spring.library.exceptions.EntityNotFoundException;
 import java.util.Collection;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.acls.model.MutableAclService;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -12,6 +15,8 @@ import org.springframework.stereotype.Service;
 public class AuthorServiceImpl implements AuthorService {
 
   private final AuthorDao authorDao;
+  private final BookDao bookDao;
+  private final MutableAclService mutableAclService;
 
   @Override
   public Collection<Author> getAll() {
@@ -20,7 +25,7 @@ public class AuthorServiceImpl implements AuthorService {
 
   @Override
   public Author create(Author author) {
-    return authorDao.save(author);
+    return save(author);
   }
 
   @Override
@@ -37,11 +42,23 @@ public class AuthorServiceImpl implements AuthorService {
 
   @Override
   public void update(Author author) {
-    authorDao.save(author);
+    save(author);
+  }
+
+  private Author save(Author author) {
+    if (authorDao.existsByName(author.getName())) {
+      throw new ConsistencyException("Cannot create author. There is another author with that name in database!");
+    } else {
+      return authorDao.save(author);
+    }
   }
 
   @Override
   public void deleteById(String id) {
-    authorDao.deleteById(id);
+    if (bookDao.existsByAuthorId(id)) {
+      throw new ConsistencyException("Cannot delete author. There are book(s) with that author in database!");
+    } else {
+      authorDao.deleteById(id);
+    }
   }
 }
