+++ homework_2/src/main/java/com/study/spring/library/service/GenreServiceImpl.java
@@ -1,7 +1,9 @@
 package com.study.spring.library.service;
 
+import com.study.spring.library.dao.BookDao;
 import com.study.spring.library.dao.GenreDao;
 import com.study.spring.library.domain.Genre;
+import com.study.spring.library.exceptions.ConsistencyException;
 import com.study.spring.library.exceptions.EntityNotFoundException;
 import java.util.Collection;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +13,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class GenreServiceImpl implements GenreService {
 
+  private final BookDao bookDao;
   private final GenreDao genreDao;
 
   @Override
@@ -20,7 +23,7 @@ public class GenreServiceImpl implements GenreService {
 
   @Override
   public Genre create(Genre genre) {
-    return genreDao.save(genre);
+    return save(genre);
   }
 
   @Override
@@ -37,11 +40,23 @@ public class GenreServiceImpl implements GenreService {
 
   @Override
   public void update(Genre genre) {
-    genreDao.save(genre);
+    save(genre);
+  }
+
+  private Genre save(Genre genre) {
+    if (genreDao.existsByName(genre.getName())) {
+      throw new ConsistencyException("Cannot create genre. There is another genre with that name in database!");
+    } else {
+      return genreDao.save(genre);
+    }
   }
 
   @Override
   public void deleteById(String id) {
-    genreDao.deleteById(id);
+    if (bookDao.existsByGenreId(id)) {
+      throw new ConsistencyException("Cannot delete genre. There are books with that genre in database!");
+    } else {
+      genreDao.deleteById(id);
+    }
   }
 }
